+++ packages/core/src/di/index.ts
@@ -18,7 +18,9 @@ export {ɵɵdefineInjectable, defineInjectable, ɵɵdefineInjector, InjectableTy
 export {forwardRef, resolveForwardRef, ForwardRefFn} from './forward_ref';
 export {Injectable, InjectableDecorator, InjectableProvider} from './injectable';
 export {Injector} from './injector';
-export {EnvironmentInjector, INJECTOR_INITIALIZER} from '../di/r3_injector';
+export {EnvironmentInjector} from './r3_injector';
+export {importProvidersFrom} from './provider_collection';
+export {INJECTOR_INITIALIZER} from './initializer_token';
 export {ProviderToken} from './provider_token';
 export {ɵɵinject, inject, ɵɵinvalidFactoryDep} from './injector_compatibility';
 export {INJECTOR} from './injector_token';
