+++ packages/core/src/di/r3_injector.ts
@@ -13,28 +13,25 @@ import {OnDestroy} from '../interface/lifecycle_hooks';
 import {Type} from '../interface/type';
 import {FactoryFn, getFactoryDef} from '../render3/definition_factory';
 import {throwCyclicDependencyError, throwInvalidProviderError, throwMixedMultiProviderError} from '../render3/errors_di';
-import {deepForEach, flatten, newArray} from '../util/array_utils';
+import {flatten, newArray} from '../util/array_utils';
 import {EMPTY_ARRAY} from '../util/empty';
 import {stringify} from '../util/stringify';
 
 import {resolveForwardRef} from './forward_ref';
+import {INJECTOR_INITIALIZER} from './initializer_token';
 import {setInjectImplementation} from './inject_switch';
 import {InjectionToken} from './injection_token';
 import {Injector} from './injector';
-import {catchInjectorError, inject, injectArgs, NG_TEMP_TOKEN_PATH, setCurrentInjector, THROW_IF_NOT_FOUND, USE_VALUE, ɵɵinject} from './injector_compatibility';
+import {catchInjectorError, injectArgs, NG_TEMP_TOKEN_PATH, setCurrentInjector, THROW_IF_NOT_FOUND, ɵɵinject} from './injector_compatibility';
 import {INJECTOR} from './injector_token';
-import {getInheritedInjectableDef, getInjectableDef, getInjectorDef, InjectorType, InjectorTypeWithProviders, ɵɵInjectableDeclaration} from './interface/defs';
+import {getInheritedInjectableDef, getInjectableDef, InjectorType, ɵɵInjectableDeclaration} from './interface/defs';
 import {InjectFlags} from './interface/injector';
-import {ClassProvider, ConstructorProvider, ExistingProvider, FactoryProvider, Provider, StaticClassProvider, StaticProvider, TypeProvider, ValueProvider} from './interface/provider';
+import {ClassProvider, ConstructorProvider, Provider, StaticClassProvider} from './interface/provider';
+import {INJECTOR_DEF_TYPES} from './internal_tokens';
 import {NullInjector} from './null_injector';
+import {importProvidersFrom, isExistingProvider, isFactoryProvider, isTypeProvider, isValueProvider, SingleProvider} from './provider_collection';
 import {ProviderToken} from './provider_token';
 import {INJECTOR_SCOPE, InjectorScope} from './scope';
-
-/**
- * Internal type for a single provider in a deep provider array.
- */
-type SingleProvider = TypeProvider|ValueProvider|ClassProvider|ConstructorProvider|ExistingProvider|
-    FactoryProvider|StaticClassProvider;
 
 /**
  * Marker which indicates that a value has not yet been created from the factory function.
@@ -49,16 +46,6 @@ const NOT_YET = {};
  * a circular dependency among the providers.
  */
 const CIRCULAR = {};
-
-/**
- * A multi-provider token for initialization functions that will run upon construction of a
- * non-view injector.
- *
- * @publicApi
- */
-export const INJECTOR_INITIALIZER = new InjectionToken<() => void>('INJECTOR_INITIALIZER');
-
-const INJECTOR_DEF_TYPES = new InjectionToken<Type<unknown>>('INJECTOR_DEF_TYPES');
 
 /**
  * A lazily initialized NullInjector.
@@ -83,174 +70,6 @@ interface Record<T> {
 }
 
 /**
- * Create a new `Injector` which is configured using a `defType` of `InjectorType<any>`s.
- *
- * @publicApi
- */
-export function createInjector(
-    defType: /* InjectorType<any> */ any, parent: Injector|null = null,
-    additionalProviders: StaticProvider[]|null = null, name?: string): Injector {
-  const injector =
-      createInjectorWithoutInjectorInstances(defType, parent, additionalProviders, name);
-  injector.resolveInjectorInitializers();
-  return injector;
-}
-
-/**
- * Creates a new injector without eagerly resolving its injector types. Can be used in places
- * where resolving the injector types immediately can lead to an infinite loop. The injector types
- * should be resolved at a later point by calling `_resolveInjectorDefTypes`.
- */
-export function createInjectorWithoutInjectorInstances(
-    defType: /* InjectorType<any> */ any, parent: Injector|null = null,
-    additionalProviders: StaticProvider[]|null = null, name?: string,
-    scopes = new Set<InjectorScope>()): R3Injector {
-  const providers = [
-    ...flatten(additionalProviders || EMPTY_ARRAY),
-    ...importProvidersFrom(defType),
-  ];
-  name = name || (typeof defType === 'object' ? undefined : stringify(defType));
-
-  return new R3Injector(providers, parent || getNullInjector(), name || null, scopes);
-}
-
-/**
- * The logic visits an `InjectorType` or `InjectorTypeWithProviders` and all of its transitive
- * providers and invokes specified callbacks when:
- * - an injector type is visited (typically an NgModule)
- * - a provider is visited
- *
- * If an `InjectorTypeWithProviders` that declares providers besides the type is specified,
- * the function will return "true" to indicate that the providers of the type definition need
- * to be processed. This allows us to process providers of injector types after all imports of
- * an injector definition are processed. (following View Engine semantics: see FW-1349)
- */
-export function walkProviderTree(
-    container: InjectorType<unknown>|InjectorTypeWithProviders<unknown>,
-    providersOut: SingleProvider[], parents: InjectorType<unknown>[],
-    dedup: Set<Type<unknown>>): container is InjectorTypeWithProviders<unknown> {
-  container = resolveForwardRef(container);
-  if (!container) return false;
-
-  // Either the defOrWrappedDef is an InjectorType (with injector def) or an
-  // InjectorDefTypeWithProviders (aka ModuleWithProviders). Detecting either is a megamorphic
-  // read, so care is taken to only do the read once.
-
-  // First attempt to read the injector def (`ɵinj`).
-  let def = getInjectorDef(container);
-
-  // If that's not present, then attempt to read ngModule from the InjectorDefTypeWithProviders.
-  const ngModule =
-      (def == null) && (container as InjectorTypeWithProviders<any>).ngModule || undefined;
-
-  // Determine the InjectorType. In the case where `defOrWrappedDef` is an `InjectorType`,
-  // then this is easy. In the case of an InjectorDefTypeWithProviders, then the definition type
-  // is the `ngModule`.
-  const defType: InjectorType<any> =
-      (ngModule === undefined) ? (container as InjectorType<any>) : ngModule;
-
-  // Check for circular dependencies.
-  if (ngDevMode && parents.indexOf(defType) !== -1) {
-    const defName = stringify(defType);
-    const path = parents.map(stringify);
-    throwCyclicDependencyError(defName, path);
-  }
-
-  // Check for multiple imports of the same module
-  const isDuplicate = dedup.has(defType);
-
-  // Finally, if defOrWrappedType was an `InjectorDefTypeWithProviders`, then the actual
-  // `InjectorDef` is on its `ngModule`.
-  if (ngModule !== undefined) {
-    def = getInjectorDef(ngModule);
-  }
-
-  // If no definition was found, it might be from exports. Remove it.
-  if (def == null) {
-    return false;
-  }
-
-  // Add providers in the same way that @NgModule resolution did:
-
-  // First, include providers from any imports.
-  if (def.imports != null && !isDuplicate) {
-    // Before processing defType's imports, add it to the set of parents. This way, if it ends
-    // up deeply importing itself, this can be detected.
-    ngDevMode && parents.push(defType);
-    // Add it to the set of dedups. This way we can detect multiple imports of the same module
-    dedup.add(defType);
-
-    let importTypesWithProviders: (InjectorTypeWithProviders<any>[])|undefined;
-    try {
-      deepForEach(def.imports, imported => {
-        if (walkProviderTree(imported, providersOut, parents, dedup)) {
-          if (importTypesWithProviders === undefined) importTypesWithProviders = [];
-          // If the processed import is an injector type with providers, we store it in the
-          // list of import types with providers, so that we can process those afterwards.
-          importTypesWithProviders.push(imported);
-        }
-      });
-    } finally {
-      // Remove it from the parents set when finished.
-      ngDevMode && parents.pop();
-    }
-
-    // Imports which are declared with providers (TypeWithProviders) need to be processed
-    // after all imported modules are processed. This is similar to how View Engine
-    // processes/merges module imports in the metadata resolver. See: FW-1349.
-    if (importTypesWithProviders !== undefined) {
-      for (let i = 0; i < importTypesWithProviders.length; i++) {
-        const {ngModule, providers} = importTypesWithProviders[i];
-        deepForEach(providers!, provider => {
-          validateProvider(provider, providers || EMPTY_ARRAY, ngModule);
-          providersOut.push(provider);
-        });
-      }
-    }
-  }
-  // Track the InjectorType and add a provider for it.
-  // It's important that this is done after the def's imports.
-  const factory = getFactoryDef(defType) || (() => new defType());
-
-  // Provider to create `defType` using its factory.
-  providersOut.push({
-    provide: defType,
-    useFactory: factory,
-    deps: EMPTY_ARRAY,
-  });
-
-  providersOut.push({
-    provide: INJECTOR_DEF_TYPES,
-    useValue: defType,
-    multi: true,
-  });
-
-  // Provider to eagerly instantiate `defType` via `INJECTOR_INITIALIZER`.
-  providersOut.push({
-    provide: INJECTOR_INITIALIZER,
-    useValue: () => inject(defType),
-    multi: true,
-  });
-
-  // Next, include providers listed on the definition itself.
-  const defProviders = def.providers;
-  if (defProviders != null && !isDuplicate) {
-    const injectorType = container as InjectorType<any>;
-    deepForEach(defProviders, provider => {
-      // TODO: fix cast
-      validateProvider(provider, defProviders as any[], injectorType);
-      providersOut.push(provider);
-    });
-  }
-
-  return (
-      ngModule !== undefined &&
-      (container as InjectorTypeWithProviders<any>).providers !== undefined);
-}
-
-
-
-/**
  * An `Injector` that's part of the environment injector hierarchy, which exists outside of the
  * component tree.
  */
@@ -273,20 +92,6 @@ export abstract class EnvironmentInjector implements Injector {
    * @internal
    */
   abstract onDestroy(callback: () => void): void;
-}
-
-/**
- * Collects providers from all NgModules, including transitively imported ones.
- *
- * @returns The list of collected providers from the specified list of NgModules.
- * @publicApi
- */
-export function importProvidersFrom(...injectorTypes: Array<Type<unknown>>): Provider[] {
-  const providers: SingleProvider[] = [];
-  deepForEach(
-      injectorTypes,
-      injectorDef => walkProviderTree(injectorDef as InjectorType<any>, providers, [], new Set()));
-  return providers;
 }
 
 export class R3Injector extends EnvironmentInjector {
@@ -633,26 +438,6 @@ function makeRecord<T>(
   };
 }
 
-function isValueProvider(value: SingleProvider): value is ValueProvider {
-  return value !== null && typeof value == 'object' && USE_VALUE in value;
-}
-
-function isExistingProvider(value: SingleProvider): value is ExistingProvider {
-  return !!(value && (value as ExistingProvider).useExisting);
-}
-
-function isFactoryProvider(value: SingleProvider): value is FactoryProvider {
-  return !!(value && (value as FactoryProvider).useFactory);
-}
-
-export function isTypeProvider(value: SingleProvider): value is TypeProvider {
-  return typeof value === 'function';
-}
-
-export function isClassProvider(value: SingleProvider): value is ClassProvider {
-  return !!(value as StaticClassProvider | ClassProvider).useClass;
-}
-
 function hasDeps(value: ClassProvider|ConstructorProvider|
                  StaticClassProvider): value is ClassProvider&{deps: any[]} {
   return !!(value as any).deps;
@@ -666,19 +451,4 @@ function hasOnDestroy(value: any): value is OnDestroy {
 function couldBeInjectableType(value: any): value is ProviderToken<any> {
   return (typeof value === 'function') ||
       (typeof value === 'object' && value instanceof InjectionToken);
-}
-
-function validateProvider(
-    provider: SingleProvider, providers: SingleProvider[], containerType: Type<unknown>): void {
-  if (isTypeProvider(provider) || isValueProvider(provider) || isFactoryProvider(provider) ||
-      isExistingProvider(provider)) {
-    return;
-  }
-
-  // Here we expect the provider to be a `useClass` provider (by elimination).
-  const classRef = resolveForwardRef(
-      provider && ((provider as StaticClassProvider | ClassProvider).useClass || provider.provide));
-  if (ngDevMode && !classRef) {
-    throwInvalidProviderError(containerType, providers, provider);
-  }
 }
