+++ packages/core/test/bundling/injection/bundle.golden_symbols.json
@@ -108,9 +108,6 @@
     "name": "importProvidersFrom"
   },
   {
-    "name": "inject"
-  },
-  {
     "name": "injectArgs"
   },
   {
