+++ packages/core/test/bundling/router/bundle.golden_symbols.json
@@ -1854,7 +1854,7 @@
     "name": "stringify"
   },
   {
-    "name": "stringify9"
+    "name": "stringify10"
   },
   {
     "name": "stringifyCSSSelector"
